requirements.txt
aiosqlite==0.20.0
alembic==1.13.3
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
async-timeout==4.0.3
asyncpg==0.30.0
bcrypt==4.2.0
billiard==4.2.1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.115.0
filelock==3.16.1
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
kombu==5.4.2
Mako==1.3.5
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.4.2
packaging==24.1
passlib==1.7.4
pillow==11.0.0
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2==2.9.9
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
sympy==1.13.1
tokenizers==0.20.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.31.0
vine==5.1.0
wcwidth==0.2.13
